[ADD] moving the API serializers into the API application; adding the serializers sub-package to the API interface application
