+++ src/interfaces/api/serializers/__init__.py
@@ -1,0 +1,28 @@
+from .datasets import (OrganismBindingSitesDetailSerializer,
+                       OrganismBindingSitesListSerializer,
+                       RegulatorBindingSitesDetailSerializer,
+                       RegulatorBindingSitesListSerializer,
+                       TRNDetailSerializer,
+                       TRNListSerializer)
+from .effector import (EffectorDetailSerializer,
+                       EffectorListSerializer)
+from .evidence import (EvidenceDetailSerializer,
+                       EvidenceListSerializer)
+from .gene import (GeneDetailSerializer,
+                   GeneListSerializer)
+from .operon import (OperonDetailSerializer,
+                     OperonListSerializer)
+from .organism import (OrganismDetailSerializer,
+                       OrganismListSerializer)
+from .pathway import (PathwayDetailSerializer,
+                      PathwayListSerializer)
+from .publication import (PublicationDetailSerializer,
+                          PublicationListSerializer)
+from .regulator import (RegulatorDetailSerializer,
+                        RegulatorListSerializer)
+from .regulatory_family import (RegulatoryFamilyDetailSerializer,
+                                RegulatoryFamilyListSerializer)
+from .regulatory_interaction import (RegulatoryInteractionDetailSerializer,
+                                     RegulatoryInteractionListSerializer)
+from .tfbs import (TFBSDetailSerializer,
+                   TFBSListSerializer)
