+++ src/interfaces/api/serializers/operon.py
@@ -2,10 +2,10 @@ from rest_framework import serializers
 
 from constants import help_text, choices
 from data import Operon
-from interfaces.serializers.base import BaseSerializer, URLField
-from interfaces.serializers.gene import GeneField
-from interfaces.serializers.relationships import (SourceRelationshipSerializer, SourceField,
-                                                  RelationshipSerializer)
+from interfaces.api.serializers.fields import GeneField
+from interfaces.serializers.base import BaseSerializer
+from interfaces.serializers.fields import URLField, SourceField
+from interfaces.serializers.relationship import SourceRelationshipSerializer, RelationshipSerializer
 
 
 class OperonListSerializer(BaseSerializer):
