+++ src/interfaces/api/serializers/organism.py
@@ -2,30 +2,9 @@ from rest_framework import serializers
 
 from constants import help_text
 from data import Organism
-from interfaces.serializers.base import BaseSerializer, URLField, NestedField
-from interfaces.serializers.relationships import (RelationshipSerializer, SourceRelationshipSerializer,
-                                                  SourceField)
-
-
-class OrganismsSerializer(BaseSerializer):
-    model = Organism
-
-    # properties
-    name = serializers.CharField(required=True, max_length=200, help_text=help_text.organism_name)
-    ncbi_taxonomy = serializers.IntegerField(required=False, min_value=0, help_text=help_text.ncbi_taxonomy)
-    species = serializers.CharField(required=False, max_length=150, help_text=help_text.species)
-    strain = serializers.CharField(required=False, max_length=150, help_text=help_text.strain)
-    refseq_accession = serializers.CharField(required=False, max_length=50,
-                                             help_text=help_text.refseq_accession)
-    genbank_accession = serializers.CharField(required=False, max_length=50,
-                                              help_text=help_text.genbank_accession)
-    assembly_accession = serializers.CharField(required=False, max_length=50,
-                                               help_text=help_text.assembly_accession)
-
-    url = URLField(read_only=True,
-                   view_name='organism',
-                   lookup_field='protrend_id',
-                   lookup_url_kwarg='protrend_id')
+from interfaces.serializers.base import BaseSerializer
+from interfaces.serializers.fields import SourceField, URLField
+from interfaces.serializers.relationship import RelationshipSerializer, SourceRelationshipSerializer
 
 
 class OrganismListSerializer(BaseSerializer):
@@ -94,10 +73,3 @@ class OrganismDetailSerializer(OrganismListSerializer):
                                                         view_name='interactions-detail',
                                                         lookup_field='protrend_id',
                                                         lookup_url_kwarg='protrend_id'))
-
-
-class OrganismField(NestedField):
-    # properties
-    protrend_id = serializers.CharField(read_only=True, help_text=help_text.protrend_id)
-    name = serializers.CharField(read_only=True, max_length=200, help_text=help_text.organism_name)
-    ncbi_taxonomy = serializers.IntegerField(read_only=True, min_value=0, help_text=help_text.ncbi_taxonomy)
